requirements.txt
biopython

.